+++ DataFormats/Scouting/interface/Run3ScoutingEERecHit.h
@@ -9,20 +9,20 @@
 
 class Run3ScoutingEERecHit {
 public:
-  Run3ScoutingEERecHit(float energy, int ix, int iy, bool positiveZ)
+  Run3ScoutingEERecHit(float energy, uint8_t ix, uint8_t iy, bool positiveZ)
       : energy_{energy}, ix_{ix}, iy_{iy}, positiveZ_{positiveZ} {}
 
   Run3ScoutingEERecHit() : energy_{0}, ix_{0}, iy_{0}, positiveZ_{false} {}
 
   float energy() const { return energy_; }
-  int ix() const { return ix_; }
-  int iy() const { return iy_; }
+  uint8_t ix() const { return ix_; }
+  uint8_t iy() const { return iy_; }
   bool positiveZ() { return positiveZ_; }
 
 private:
   float energy_;
-  int ix_;
-  int iy_;
+  uint8_t ix_;
+  uint8_t iy_;
   bool positiveZ_;
 };
